Изменения из dz3
