+++ .gitignore
@@ -1,6 +1,4 @@
 node_modules/
 .idea/
 simple-chat/build
-simple-chat/
 simple-chat/src/icons
-.github
