--- package.json
@@ -1,28 +1,0 @@
-{
-  "name": "homework-vk-1",
-  "version": "1.0.0",
-  "main": "convertBytesToHuman.js",
-  "scripts": {
-    "test": "jest"
-  },
-  "jest": {
-    "transform": {
-      "^.+\\.jsx?$": "babel-jest"
-    }
-  },
-  "babel": {
-    "presets": [
-      "@babel/preset-env"
-    ]
-  },
-  "author": "я",
-  "license": "ISC",
-  "description": "",
-  "devDependencies": {
-    "@babel/core": "^7.25.2",
-    "@babel/preset-env": "^7.25.4",
-    "babel-jest": "^29.7.0",
-    "file-loader": "^6.2.0",
-    "jest": "^29.7.0"
-  }
-}
